// src/app/struct/types.ts
export class TRouteData {
    caption: string;
    href: string;
    comp: any;
    hideMenu: boolean;
}
export class TUserLogin {
    username: string;
    userpass: string;
}
export class TUser {
    username: string;
    userpass: string;
    firstname: string;
    lastname: string;
}
export class TServerResponse {
    messages: any[];
    status: number;
}